Refactor a CustomMapEvents component for handling map events
